improv: add option to not cleanup devCopyDir

+++ src/serve.ts
@@ -7,18 +7,21 @@ import { copyAll, outputCopyLog } from './utils'
 export const servePlugin = ({
   devCopyDir,
   targets,
-  flatten
+  flatten,
+  cleanup
 }: ResolvedViteStaticCopyOptions): Plugin => {
   return {
     name: 'vite-plugin-static-copy:serve',
     apply: 'serve',
     async buildStart() {
-      // copy again
-      try {
-        await fs.rm(devCopyDir, { force: true, recursive: true })
-        // eslint-disable-next-line no-empty
-      } catch {}
-      await fs.mkdir(devCopyDir, { recursive: true })
+      if (cleanup) {
+        // copy again
+        try {
+          await fs.rm(devCopyDir, { force: true, recursive: true })
+          // eslint-disable-next-line no-empty
+        } catch {}
+        await fs.mkdir(devCopyDir, { recursive: true })
+      }
 
       const copyCount = await copyAll(devCopyDir, targets, flatten)
       outputCopyLog(copyCount)
